Makefile generator fixes

+++ src/scripts/MakefileGenerator/MakGenGCCMakefiles.js
@@ -134,7 +134,7 @@ function GnuCxx( F, P )
    P.validateGnuCxxBuild();
 
    var buildDirectory = P.gccBuildDirectory( F.baseDirectory );
-   if ( P.cleanUpPreviousBuilds )
+   if ( P.cleanUpPreviousBuilds && !P.gccDebug )
       if ( File.directoryExists( buildDirectory ) )
          removeDirectory( buildDirectory );
 
@@ -225,8 +225,8 @@ function GnuCxx( F, P )
     * Without this, we'd be propagating our own build directories everywhere!
     */
    f.outTextLn( "OBJ_DIR=\"" + objectDirectory.replace(
-                                 RegExp( '^' + PCLSRCDIR ), "$(PCLSRCDIR)" ).replace(
-                                    RegExp( '^' + PCLDIR ), "$(PCLDIR)" ) + "\"" );
+                                 RegExp( '^' + escapeForRegExp(PCLSRCDIR) ), "$(PCLSRCDIR)" ).replace(
+                                    RegExp( '^' + escapeForRegExp(PCLDIR) ), "$(PCLDIR)" ) + "\"" );
    f.outTextLn( '' );
 
    f.outTextLn( ".PHONY: all" );
